Made Zedram::Parser use the newer perl6 file reading paradigms (like slurp and lines).
Fixed a bug in the change_parser_using_token sub that tried to convert NONE into null.  I, when writing it, forgot that the GrammarDelimiter will be passed with the value as the starting char.  Because it is useful to see if a value was passed successfully, I decided to leave in the grammar delimiter (so people who extend the language can tell if a value will be passed) and therefore made that particular rule only look for NONE at the end.

Also added in three examples (they are present on the wiki) and made the compiler.pl use those files.